WIP: Recreate Swapchain
Need to get the event from CrossWindow
CrossWindow is purely pollbased -> Callback Vectors, add register functions that add to an array of specific functions, which will call with the window number and parameters, exactly like glfw did
Then hookup from the engine to the TriggerResizeSwapchain with the correct size, just change the bool and the area so that next time it is resized

+++ engine/mfly_window/mfly_window.hpp
@@ -53,7 +53,7 @@ namespace mfly
             uint8_t id = 0;
         };
         
-        void* getGAPISurface(void* gapi_instance, uint16_t window_handle);
+        void* getGAPISurface(void* gapi_instance, uint32_t window_handle);
         
     };
 };
